Job_matching_Screened/requirements.txt
pandas
openpyxl
gspread
oauth2client
customtkinter
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
tk
PyQt6


